+++ UVtools.Core/Extensions/SizeExtensions.cs
@@ -38,5 +38,36 @@ namespace UVtools.Core.Extensions
 
         public static Size Inflate(this Size size, int pixels) => new (size.Width + pixels, size.Height + pixels);
         public static Size Inflate(this Size size, int width, int height) => new (size.Width + width, size.Height + height);
+
+        public static int Area(this Rectangle rect)
+        {
+            return rect.Width * rect.Height;
+        }
+
+        public static int Area(this Size size)
+        {
+            return size.Width * size.Height;
+        }
+
+        public static int Max(this Size size)
+        {
+            return Math.Max(size.Width, size.Height);
+        }
+
+        public static float Area(this RectangleF rect)
+        {
+            return rect.Width * rect.Height;
+        }
+
+        public static float Area(this SizeF size)
+        {
+            return size.Width * size.Height;
+        }
+
+        public static float Max(this SizeF size)
+        {
+            return Math.Max(size.Width, size.Height);
+        }
+
     }
 }