+++ UVtools.WPF/MainWindow.Information.cs
@@ -344,20 +344,25 @@ namespace UVtools.WPF
             //CurrentLayerProperties.Add(new StringTag(nameof(layer.NonZeroPixelCount), layer.NonZeroPixelCount.ToString()));
             CurrentLayerProperties.Add(new StringTag(nameof(layer.ExposureTime), $"{layer.ExposureTime:F2}s"));
 
-            if (ReferenceEquals(SlicerFile.PrintParameterPerLayerModifiers, null)) return;
+            if (SlicerFile.PrintParameterPerLayerModifiers is not null)
+            {
 
-            if (SlicerFile.PrintParameterPerLayerModifiers.Contains(FileFormat.PrintParameterModifier.LiftHeight))
-                CurrentLayerProperties.Add(new StringTag(nameof(layer.LiftHeight),
-                    $"{layer.LiftHeight.ToString(CultureInfo.InvariantCulture)}mm @ {layer.LiftSpeed.ToString(CultureInfo.InvariantCulture)}mm/min"));
-            //CurrentLayerProperties.Adnew StringTagg>(nameof(layer.LiftSpeed), $"{layer.LiftSpeed.ToString(CultureInfo.InvariantCulture)}mm/min"));
-            if (SlicerFile.PrintParameterPerLayerModifiers.Contains(FileFormat.PrintParameterModifier.RetractSpeed))
-                CurrentLayerProperties.Add(new StringTag(nameof(layer.RetractSpeed),
-                    $"{layer.RetractSpeed.ToString(CultureInfo.InvariantCulture)}mm/min"));
-            if (SlicerFile.PrintParameterPerLayerModifiers.Contains(FileFormat.PrintParameterModifier.LightOffDelay))
-                CurrentLayerProperties.Add(new StringTag(nameof(layer.LightOffDelay),
-                    $"{layer.LightOffDelay.ToString(CultureInfo.InvariantCulture)}s"));
-            if (SlicerFile.PrintParameterPerLayerModifiers.Contains(FileFormat.PrintParameterModifier.LightPWM))
-                CurrentLayerProperties.Add(new StringTag(nameof(layer.LightPWM), layer.LightPWM.ToString()));
+                if (SlicerFile.PrintParameterPerLayerModifiers.Contains(FileFormat.PrintParameterModifier.LiftHeight))
+                    CurrentLayerProperties.Add(new StringTag(nameof(layer.LiftHeight),
+                        $"{layer.LiftHeight.ToString(CultureInfo.InvariantCulture)}mm @ {layer.LiftSpeed.ToString(CultureInfo.InvariantCulture)}mm/min"));
+                //CurrentLayerProperties.Adnew StringTagg>(nameof(layer.LiftSpeed), $"{layer.LiftSpeed.ToString(CultureInfo.InvariantCulture)}mm/min"));
+                if (SlicerFile.PrintParameterPerLayerModifiers.Contains(FileFormat.PrintParameterModifier.RetractSpeed))
+                    CurrentLayerProperties.Add(new StringTag(nameof(layer.RetractSpeed),
+                        $"{layer.RetractSpeed}mm/min"));
+                if (SlicerFile.PrintParameterPerLayerModifiers.Contains(FileFormat.PrintParameterModifier.LightOffDelay)
+                )
+                    CurrentLayerProperties.Add(new StringTag(nameof(layer.LightOffDelay),
+                        $"{layer.LightOffDelay}s"));
+                if (SlicerFile.PrintParameterPerLayerModifiers.Contains(FileFormat.PrintParameterModifier.LightPWM))
+                    CurrentLayerProperties.Add(new StringTag(nameof(layer.LightPWM), layer.LightPWM.ToString()));
+            }
+
+            CurrentLayerProperties.Add(new StringTag(nameof(layer.MaterialMilliliters), $"{layer.MaterialMilliliters:F4}ml"));
         }
         #endregion
     }